New filter settings, ADD_filters option
Main changes:

- Dictionary ADDfilters[] was added, inside the filters[] dictionary.
In ADDfilters[] the user can enter the names, files and other
properties of the filters he/she wants to add. These will be then read
out in FILTER_AGNfitter.py
- Function MODELS_settings() (with dictionary models []) was added,
which will be further implemented soon.

Minor changes:

- cat['dict_path'] was removed. The name of the dictionary of models
will be now automatically created as a combination of the names of the
filterset (specified by the user in filters['filterset’]) and the
modelset (specified by the user in models['modelset’] ). This should
give less issues with the overwriting of dictionaries.
-FITS handling was already implemented, but comments in this file were
not updated till now.