New model system - new functions in DICTIONARIES_AGNfitter.py
MAIN CHANGES:

- function construct_dictionaryarray_filtered is now independent of
specific parameters but accepts general lists and arrays of the
parameters for each model, which were previously  defined in
MODEL_AGNfitter.py

-function dictkey_arrays constructs the class ‘pick_obj’, which
transports the model parameter properties throughout the code. This
class replaces the functions pick_COMPONENT_template, which used to be
in MODEL_AGNfitter.py.

- function filtering_models was moved here from  MODEL_AGNfitter.py, as
it is only used in routines of this script.