New models system - generalised to any sets of parameters
MAIN CHANGES:

- P dictionary is generalised for any   parameter sets, since it
automatically reads the parameter names defined for each component in
MODEL_AGNfitter.py. It also sets the minimum and maximum values for the
parameters which are used as limits for the flat prior.

- ymodel is also generalised to any sets of parameters. It uses the P
dictionary for getting the parameter information and the class pick_obj
as the interface between MCMC and the model space.

MINOR CHANGES:
- P dictionary is a python dictionary (not an instance of the class
addict any longer), thus contains keys defined as P[‘key’].